coda: allow large GOP sizes and GOP size 0 on CODA960
CodaDx6 only accepts GOP sizes up to 60 frames, CODA960 accepts up to 99
frames. If we disable automatic I frame generation altogether (by setting
GOP size to 0), we can produce arbitrarily large GOP sizes as soon as we
can force I frames from userspace with an IDR mechanism.

+++ drivers/media/platform/coda/coda.h
@@ -121,7 +121,7 @@ struct coda_params {
 	u8			h264_level_idc;
 	u8			mpeg4_intra_qp;
 	u8			mpeg4_inter_qp;
-	u8			gop_size;
+	int			gop_size;
 	int			intra_refresh;
 	u8			jpeg_quality;
 	u8			jpeg_restart_interval;
